Fix calendar TBD: match short month format in extractTimesFromText
datesText uses 'Jun' (short) but label lookup used 'June' (long), causing
find() to always fail and return TBD.

+++ account.js
@@ -957,7 +957,7 @@ function fmt12Acc(t) {
 function extractTimesFromText(datesText, iso) {
     const lines = (datesText || '').split('\n').map(l => l.trim()).filter(Boolean);
     const d = new Date(iso + 'T12:00:00');
-    const label = d.toLocaleDateString('en-US', { month:'long', day:'numeric', year:'numeric' });
+    const label = d.toLocaleDateString('en-US', { month:'short', day:'numeric', year:'numeric' });
     const line = lines.find(l => l.includes(label));
     if (!line) return ['TBD'];
     const m = line.match(/:\s*(.+)$/);
